+++ src/mame/tatsumi/tatsumi.cpp
@@ -829,13 +829,13 @@ static const gfx_layout roundup5_vramlayout =
 };
 
 static GFXDECODE_START( gfx_apache3_sprites )
-	GFXDECODE_ENTRY( "sprites_l", 0, spritelayout,    0, 256)
-	GFXDECODE_ENTRY( "sprites_h", 0, spritelayout,    0, 256)
+	GFXDECODE_ENTRY( "rotatingsprites:sprites_l", 0, spritelayout,    0, 256)
+	GFXDECODE_ENTRY( "rotatingsprites:sprites_h", 0, spritelayout,    0, 256)
 GFXDECODE_END
 
 static GFXDECODE_START( gfx_cyclwarr_sprites )
-	GFXDECODE_ENTRY( "sprites_l", 0, spritelayout,  0, 512)
-	GFXDECODE_ENTRY( "sprites_h", 0, spritelayout,  0, 512)
+	GFXDECODE_ENTRY( "rotatingsprites:sprites_l", 0, spritelayout,  0, 512)
+	GFXDECODE_ENTRY( "rotatingsprites:sprites_h", 0, spritelayout,  0, 512)
 GFXDECODE_END
 
 static GFXDECODE_START( gfx_apache3 )
@@ -907,6 +907,11 @@ void apache3_state::apache3(machine_config &config)
 	screen.set_raw(CLOCK_2 / 8, 400, 0, 320, 272, 0, 240); // TODO: Hook up CRTC
 	screen.set_screen_update(FUNC(apache3_state::screen_update_apache3));
 
+	TATSUMI_ROTATING_SPRITES(config, m_rotatingsprites, 0);
+	m_rotatingsprites->set_rom_clut_offset(0x100000 - 0x800);
+	m_rotatingsprites->set_rom_clut_size(0x800);
+	m_rotatingsprites->set_sprite_palette_base(0);
+
 	GFXDECODE(config, m_spritegfxdecode, m_fakepalette, gfx_apache3_sprites);
 	GFXDECODE(config, m_gfxdecode, m_palette, gfx_apache3);
 	PALETTE(config, m_palette).set_format(palette_device::xRGB_555, 2048); // 2048 real colours
@@ -960,6 +965,11 @@ void roundup5_state::roundup5(machine_config &config)
 	screen_device &screen(SCREEN(config, "screen", SCREEN_TYPE_RASTER));
 	screen.set_raw(CLOCK_2 / 8, 400, 0, 320, 272, 0, 240); // TODO: Hook up CRTC
 	screen.set_screen_update(FUNC(roundup5_state::screen_update_roundup5));
+
+	TATSUMI_ROTATING_SPRITES(config, m_rotatingsprites, 0);
+	m_rotatingsprites->set_rom_clut_offset(0xc0000 - 0x800);
+	m_rotatingsprites->set_rom_clut_size(0x800);
+	m_rotatingsprites->set_sprite_palette_base(512);
 
 	GFXDECODE(config, m_spritegfxdecode, m_fakepalette, gfx_apache3_sprites);
 	GFXDECODE(config, m_gfxdecode, m_palette, gfx_roundup5);
@@ -1045,6 +1055,11 @@ void cyclwarr_state::cyclwarr(machine_config &config)
 	screen.set_raw(CLOCK_2 / 8, 400, 0, 320, 272, 0, 240); // TODO: Hook up CRTC
 	screen.set_screen_update(FUNC(cyclwarr_state::screen_update_cyclwarr));
 
+	TATSUMI_ROTATING_SPRITES(config, m_rotatingsprites, 0);
+	m_rotatingsprites->set_rom_clut_offset(0x100000 - 0x1000);
+	m_rotatingsprites->set_rom_clut_size(0x1000);
+	m_rotatingsprites->set_sprite_palette_base(4096);
+
 	GFXDECODE(config, m_spritegfxdecode, m_fakepalette, gfx_cyclwarr_sprites);
 	GFXDECODE(config, m_gfxdecode, m_palette, gfx_cyclwarr);
 	PALETTE(config, m_palette).set_format(palette_device::xRGB_555, 8192);
@@ -1075,6 +1090,7 @@ void cyclwarr_state::bigfight(machine_config &config)
 
 	// TODO: it's same video HW, we don't know how/where video registers are mapped
 //  subdevice<screen_device>("screen")->set_screen_update(FUNC(cyclwarr_state::screen_update_bigfight));
+	m_rotatingsprites->set_rom_clut_offset(0x200000 - 0x1000);
 
 	MCFG_VIDEO_START_OVERRIDE(cyclwarr_state, bigfight)
 
@@ -1101,7 +1117,7 @@ ROM_START( apache3 )
 	ROM_REGION( 0x10000, "sound_rom", 0 ) /* 64k code for sound V20 */
 	ROM_LOAD( "ap-27d.151",   0x00000, 0x10000, CRC(294b4d79) SHA1(2b03418a12a2aaf3919b98161d8d0ce6ae29a2bb) )
 
-	ROM_REGION( 0x100000, "sprites_l", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "ap-00c.15",   0x000000, 0x20000, CRC(ad1ddc2b) SHA1(81f64663c4892ab5fb0e2dc99513dbfee73f15b8) )
 	ROM_LOAD32_BYTE( "ap-01c.22",   0x000001, 0x20000, CRC(6286ff00) SHA1(920da4a3a441dbf54ad86c0f4fb6f47a867e9cda) )
 	ROM_LOAD32_BYTE( "ap-04c.58",   0x000002, 0x20000, CRC(dc6d55e4) SHA1(9f48f8d6aa1a329a71913139a8d5a50d95a9b9e5) )
@@ -1111,7 +1127,7 @@ ROM_START( apache3 )
 	ROM_LOAD32_BYTE( "ap-06c.71",   0x080002, 0x20000, CRC(0ea90e55) SHA1(b16d6b8be4853797507d3e5c933a9dd1d451308e) )
 	ROM_LOAD32_BYTE( "ap-07c.75",   0x080003, 0x20000, CRC(ba685543) SHA1(140a2b708d4e4de4d207fc2c4a96a5cab8639988) )
 
-	ROM_REGION( 0x100000, "sprites_h", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "ap-08c.14",   0x000000, 0x20000, CRC(6437b580) SHA1(2b2ba42add18bbec04fbcf53645a8d44b972e26a) )
 	ROM_LOAD32_BYTE( "ap-09c.21",   0x000001, 0x20000, CRC(54d18ef9) SHA1(40ebc6ea49b2a501fe843d60bec8c32d07f2d25d) )
 	ROM_LOAD32_BYTE( "ap-12c.57",   0x000002, 0x20000, CRC(f95cf5cf) SHA1(ce373c648cbf3e4863bbc3a1175efe065c75eb13) )
@@ -1150,7 +1166,7 @@ ROM_START( apache3a )
 	ROM_REGION( 0x10000, "sound_rom", 0 ) /* 64k code for sound V20 */
 	ROM_LOAD( "ap-27d.151",   0x00000, 0x10000, CRC(294b4d79) SHA1(2b03418a12a2aaf3919b98161d8d0ce6ae29a2bb) )
 
-	ROM_REGION( 0x100000, "sprites_l", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "ap-00c.15",   0x000000, 0x20000, CRC(ad1ddc2b) SHA1(81f64663c4892ab5fb0e2dc99513dbfee73f15b8) )
 	ROM_LOAD32_BYTE( "ap-01c.22",   0x000001, 0x20000, CRC(6286ff00) SHA1(920da4a3a441dbf54ad86c0f4fb6f47a867e9cda) )
 	ROM_LOAD32_BYTE( "ap-04c.58",   0x000002, 0x20000, CRC(dc6d55e4) SHA1(9f48f8d6aa1a329a71913139a8d5a50d95a9b9e5) )
@@ -1160,7 +1176,7 @@ ROM_START( apache3a )
 	ROM_LOAD32_BYTE( "ap-06c.71",   0x080002, 0x20000, CRC(0ea90e55) SHA1(b16d6b8be4853797507d3e5c933a9dd1d451308e) )
 	ROM_LOAD32_BYTE( "ap-07c.75",   0x080003, 0x20000, CRC(ba685543) SHA1(140a2b708d4e4de4d207fc2c4a96a5cab8639988) )
 
-	ROM_REGION( 0x100000, "sprites_h", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "ap-08c.14",   0x000000, 0x20000, CRC(6437b580) SHA1(2b2ba42add18bbec04fbcf53645a8d44b972e26a) )
 	ROM_LOAD32_BYTE( "ap-09c.21",   0x000001, 0x20000, CRC(54d18ef9) SHA1(40ebc6ea49b2a501fe843d60bec8c32d07f2d25d) )
 	ROM_LOAD32_BYTE( "ap-12c.57",   0x000002, 0x20000, CRC(f95cf5cf) SHA1(ce373c648cbf3e4863bbc3a1175efe065c75eb13) )
@@ -1199,7 +1215,7 @@ ROM_START( apache3b )
 	ROM_REGION( 0x10000, "sound_rom", 0 ) /* 64k code for sound V20 */
 	ROM_LOAD( "ap-27d.151",   0x00000, 0x10000, CRC(294b4d79) SHA1(2b03418a12a2aaf3919b98161d8d0ce6ae29a2bb) )
 
-	ROM_REGION( 0x100000, "sprites_l", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "ap-00c.15",   0x000000, 0x20000, CRC(ad1ddc2b) SHA1(81f64663c4892ab5fb0e2dc99513dbfee73f15b8) )
 	ROM_LOAD32_BYTE( "ap-01c.22",   0x000001, 0x20000, CRC(6286ff00) SHA1(920da4a3a441dbf54ad86c0f4fb6f47a867e9cda) )
 	ROM_LOAD32_BYTE( "ap-04c.58",   0x000002, 0x20000, CRC(dc6d55e4) SHA1(9f48f8d6aa1a329a71913139a8d5a50d95a9b9e5) )
@@ -1209,7 +1225,7 @@ ROM_START( apache3b )
 	ROM_LOAD32_BYTE( "ap-06c.71",   0x080002, 0x20000, CRC(0ea90e55) SHA1(b16d6b8be4853797507d3e5c933a9dd1d451308e) )
 	ROM_LOAD32_BYTE( "ap-07c.75",   0x080003, 0x20000, CRC(ba685543) SHA1(140a2b708d4e4de4d207fc2c4a96a5cab8639988) )
 
-	ROM_REGION( 0x100000, "sprites_h", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "ap-08c.14",   0x000000, 0x20000, CRC(6437b580) SHA1(2b2ba42add18bbec04fbcf53645a8d44b972e26a) )
 	ROM_LOAD32_BYTE( "ap-09c.21",   0x000001, 0x20000, CRC(54d18ef9) SHA1(40ebc6ea49b2a501fe843d60bec8c32d07f2d25d) )
 	ROM_LOAD32_BYTE( "ap-12c.57",   0x000002, 0x20000, CRC(f95cf5cf) SHA1(ce373c648cbf3e4863bbc3a1175efe065c75eb13) )
@@ -1248,7 +1264,7 @@ ROM_START( roundup5 )
 	ROM_REGION( 0x10000, "audiocpu", 0 ) /* 64k code for sound Z80 */
 	ROM_LOAD( "ru-28d",   0x000000, 0x10000, CRC(df36c6c5) SHA1(c046482043f6b54c55696ba3d339ffb11d78f674) )
 
-	ROM_REGION( 0x0c0000, "sprites_l", 0)
+	ROM_REGION( 0x0c0000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "ru-00b",   0x000000, 0x20000, CRC(388a0647) SHA1(e4ab43832872f44c0fe1aaede4372cc00ca7d32b) )
 	ROM_LOAD32_BYTE( "ru-02b",   0x000001, 0x20000, CRC(eff33945) SHA1(3f4c3aaa11ccf945c2f898dfdf815705d8539e21) )
 	ROM_LOAD32_BYTE( "ru-04b",   0x000002, 0x20000, CRC(40fda247) SHA1(f5fbc07fda024baedf35ac209210e94df9f15065) )
@@ -1258,7 +1274,7 @@ ROM_START( roundup5 )
 	ROM_LOAD32_BYTE( "ru-05b",   0x080002, 0x10000, CRC(23dd10e1) SHA1(f30ff1a8c7ed9bc567b901cbdd202028fffb9f80) )
 	ROM_LOAD32_BYTE( "ru-07b",   0x080003, 0x10000, CRC(bb40f46e) SHA1(da694e16d19f60a0dee47551f00f3e50b2d5dcaf) )
 
-	ROM_REGION( 0x0c0000, "sprites_h", 0)
+	ROM_REGION( 0x0c0000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "ru-08b",   0x000000, 0x20000, CRC(01729e3c) SHA1(1445287fde0b993d053aab73efafc902a6b7e2cc) )
 	ROM_LOAD32_BYTE( "ru-10b",   0x000001, 0x20000, CRC(cd2357a7) SHA1(313460a74244325ce2c659816f2b738f3dc5358a) )
 	ROM_LOAD32_BYTE( "ru-12b",   0x000002, 0x20000, CRC(ca63b1f8) SHA1(a50ef8259745dc166eb0a1b2c812ff620818a755) )
@@ -1297,7 +1313,7 @@ ROM_START( cyclwarr )
 	ROM_REGION( 0x10000, "audiocpu", 0 ) /* 64k code for sound Z80 */
 	ROM_LOAD( "cw26a.ic91", 0x000000, 0x10000, CRC(f7a70e3a) SHA1(5581633bf1f15d7f5c1e03de897d65d60f9f1e33) )
 
-	ROM_REGION( 0x100000, "sprites_l", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "cw00a.ic26", 0x000000, 0x20000, CRC(058a77f1) SHA1(93f99fcf6ce6714d76af6f6e930115516f0379d3) )
 	ROM_LOAD32_BYTE( "cw08a.ic45", 0x000001, 0x20000, CRC(f53993e7) SHA1(ef2d502ab180d2bc0bdb698c2878fdee9a2c33a8) )
 	ROM_LOAD32_BYTE( "cw02a.ic28", 0x000002, 0x20000, CRC(4dadf3cb) SHA1(e42c56e295a443cb605d48eba23a16fab3c86525) )
@@ -1307,7 +1323,7 @@ ROM_START( cyclwarr )
 	ROM_LOAD32_BYTE( "cw03a.ic29", 0x080002, 0x20000, CRC(3ca6f98e) SHA1(8526fe38d3b4c66e09049ba18651a9e7255d85d6) )
 	ROM_LOAD32_BYTE( "cw11a.ic48", 0x080003, 0x20000, CRC(5d760392) SHA1(7bbda2880af4659c267193ce10ed887a1b54a981) )
 
-	ROM_REGION( 0x100000, "sprites_h", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "cw04a.ic30", 0x000000, 0x20000, CRC(f05f594d) SHA1(80effaa517b2154c013419e0bc05fd0797b74c8d) )
 	ROM_LOAD32_BYTE( "cw12a.ic49", 0x000001, 0x20000, CRC(4ac07e8b) SHA1(f9de96fba39d5752d61b8f6be87fb605694624ed) )
 	ROM_LOAD32_BYTE( "cw06a.ic32", 0x000002, 0x20000, CRC(f628edc9) SHA1(473f7ec28000e6bf72782c1c3f4afb5e021bd430) )
@@ -1354,7 +1370,7 @@ ROM_START( cyclwarra )
 	ROM_REGION( 0x10000, "audiocpu", 0 ) /* 64k code for sound Z80 */
 	ROM_LOAD( "cw26a.ic91", 0x000000, 0x10000, CRC(f7a70e3a) SHA1(5581633bf1f15d7f5c1e03de897d65d60f9f1e33) )
 
-	ROM_REGION( 0x100000, "sprites_l", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "cw00a.ic26", 0x000000, 0x20000, CRC(058a77f1) SHA1(93f99fcf6ce6714d76af6f6e930115516f0379d3) )
 	ROM_LOAD32_BYTE( "cw08a.ic45", 0x000001, 0x20000, CRC(f53993e7) SHA1(ef2d502ab180d2bc0bdb698c2878fdee9a2c33a8) )
 	ROM_LOAD32_BYTE( "cw02a.ic28", 0x000002, 0x20000, CRC(4dadf3cb) SHA1(e42c56e295a443cb605d48eba23a16fab3c86525) )
@@ -1364,7 +1380,7 @@ ROM_START( cyclwarra )
 	ROM_LOAD32_BYTE( "cw03a.ic29", 0x080002, 0x20000, CRC(3ca6f98e) SHA1(8526fe38d3b4c66e09049ba18651a9e7255d85d6) )
 	ROM_LOAD32_BYTE( "cw11a.ic48", 0x080003, 0x20000, CRC(5d760392) SHA1(7bbda2880af4659c267193ce10ed887a1b54a981) )
 
-	ROM_REGION( 0x100000, "sprites_h", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "cw04a.ic30", 0x000000, 0x20000, CRC(f05f594d) SHA1(80effaa517b2154c013419e0bc05fd0797b74c8d) )
 	ROM_LOAD32_BYTE( "cw12a.ic49", 0x000001, 0x20000, CRC(4ac07e8b) SHA1(f9de96fba39d5752d61b8f6be87fb605694624ed) )
 	ROM_LOAD32_BYTE( "cw06a.ic32", 0x000002, 0x20000, CRC(f628edc9) SHA1(473f7ec28000e6bf72782c1c3f4afb5e021bd430) )
@@ -1411,7 +1427,7 @@ ROM_START( cyclwarrb )
 	ROM_REGION( 0x10000, "audiocpu", 0 ) /* 64k code for sound Z80 */
 	ROM_LOAD( "cw26.ic91", 0x000000, 0x10000, CRC(a6485a3a) SHA1(b4fcf541efe48b3ca32065221fe2f59476a4f96a) )
 
-	ROM_REGION( 0x100000, "sprites_l", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_l", 0)
 	ROM_LOAD32_BYTE( "cw00.ic26", 0x000000, 0x20000, CRC(ba00c582) SHA1(2cd645b828595acbe62e2f7aad037fcbdc5a543f) )
 	ROM_LOAD32_BYTE( "cw08.ic45", 0x000001, 0x20000, CRC(1583e576) SHA1(646762d1d181231090a18698378f60d09f26f49f) )
 	ROM_LOAD32_BYTE( "cw02.ic28", 0x000002, 0x20000, CRC(8376a744) SHA1(633d20199382f760adfb528f5b13730ddf9016e3) )
@@ -1421,7 +1437,7 @@ ROM_START( cyclwarrb )
 	ROM_LOAD32_BYTE( "cw03.ic29", 0x080002, 0x20000, CRC(951ed812) SHA1(b3db6b467fd626936568773367099c9abcabfab6) )
 	ROM_LOAD32_BYTE( "cw11.ic48", 0x080003, 0x20000, CRC(a7e5bf0b) SHA1(883b943d40f4516a21692beffb12514ad9301f20) )
 
-	ROM_REGION( 0x100000, "sprites_h", 0)
+	ROM_REGION( 0x100000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "cw04.ic30", 0x000000, 0x20000, CRC(890ea7b1) SHA1(737e58800aa6863aff043ba46c9cebc8ba6c1501) )
 	ROM_LOAD32_BYTE( "cw12.ic49", 0x000001, 0x20000, CRC(1587e96d) SHA1(2ffcb27d90ef29bc79d0a29f46a1d43565935a15) )
 	ROM_LOAD32_BYTE( "cw06.ic32", 0x000002, 0x20000, CRC(47decb23) SHA1(4868c01035175698cb8af7aae80627b51583213f) )
@@ -1460,13 +1476,13 @@ ROM_START( bigfight )
 	ROM_REGION( 0x10000, "audiocpu", 0 ) /* 64k code for sound Z80 */
 	ROM_LOAD( "rom20.ic91",   0x000000, 0x10000, CRC(b3add091) SHA1(8a67bfff75c13fe4d9b89d30449199200d11cea7) ) // == bf36b.ic91
 
-	ROM_REGION( 0x200000, "sprites_l", 0 )
+	ROM_REGION( 0x200000, "rotatingsprites:sprites_l", 0 )
 	ROM_LOAD32_BYTE( "rom0.ic26",   0x000000, 0x80000, CRC(a4a3c8d6) SHA1(b5365d9bc6068260c23ba9d5971c7c7d7cc07a97) )
 	ROM_LOAD32_BYTE( "rom8.ic45",   0x000001, 0x80000, CRC(220956ed) SHA1(68e0ba1e850101b4cc2778819dfa76f04d88d2d6) )
 	ROM_LOAD32_BYTE( "rom2.ic28",   0x000002, 0x80000, CRC(c4f6d243) SHA1(e23b241b5a40b332165a34e2f1bc4366973b2070) )
 	ROM_LOAD32_BYTE( "rom10.ic47",  0x000003, 0x80000, CRC(0212d472) SHA1(5549461195fd7b6b43c0174462d7fe1a1bac24e9) )
 
-	ROM_REGION( 0x200000, "sprites_h", 0 )
+	ROM_REGION( 0x200000, "rotatingsprites:sprites_h", 0 )
 	ROM_LOAD32_BYTE( "rom4.ic30",   0x000000, 0x80000, CRC(999ff7e9) SHA1(a53b06ad084722d7a52fcf01c52967f68620e609) )
 	ROM_LOAD32_BYTE( "rom12.ic49",  0x000001, 0x80000, CRC(cb4c1f0b) SHA1(32d64b78ed3d5971eb5d25be2c38e6f2c9048f74) )
 	ROM_LOAD32_BYTE( "rom6.ic32",   0x000002, 0x80000, CRC(f70e2d47) SHA1(00517b5f3b2deb6f3f3bd12df421e63884c22b2e) )
@@ -1508,7 +1524,7 @@ ROM_START( bigfightj ) // ABA-011 main board + ABA-012 daughter board
 	ROM_REGION( 0x10000, "audiocpu", 0 ) /* 64k code for sound Z80 */
 	ROM_LOAD( "bf36b.ic91", 0x000000, 0x10000, CRC(b3add091) SHA1(8a67bfff75c13fe4d9b89d30449199200d11cea7) ) // rev B
 
-	ROM_REGION( 0x200000, "sprites_l", 0 )
+	ROM_REGION( 0x200000, "rotatingsprites:sprites_l", 0 )
 	ROM_LOAD32_BYTE( "bf00d.ic26", 0x000000, 0x40000, CRC(f506d508) SHA1(86255631ac139f1b5c0f5d6e54a0858625497a1e) ) // all rev D
 	ROM_LOAD32_BYTE( "bf08d.ic45", 0x000001, 0x40000, CRC(4bf948b9) SHA1(c65b95c454d04c7e4a0cf426ac2d45ddc5e8885e) )
 	ROM_LOAD32_BYTE( "bf02d.ic28", 0x000002, 0x40000, CRC(af30acf7) SHA1(8f4778b33abb18b113d5d27c0e957d633557c988) )
@@ -1517,9 +1533,9 @@ ROM_START( bigfightj ) // ABA-011 main board + ABA-012 daughter board
 	ROM_LOAD32_BYTE( "bf09d.ic46", 0x100001, 0x20000, CRC(284837ed) SHA1(b1c130b45ff0f22985962240c47b7b01df6ac636) )
 	ROM_LOAD32_BYTE( "bf03d.ic29", 0x100002, 0x20000, CRC(2ba0398e) SHA1(ec9a29b661b18980c07a446afc89becb1ebddd57) )
 	ROM_LOAD32_BYTE( "bf11d.ic48", 0x100003, 0x20000, CRC(3f2fa72f) SHA1(4b4821b6933ea753e092f11d80bcc7698f85ccf2) )
-	ROM_COPY("sprites_l",          0x100000, 0x180000, 0x080000 )
+	ROM_COPY("rotatingsprites:sprites_l",          0x100000, 0x180000, 0x080000 )
 
-	ROM_REGION( 0x200000, "sprites_h", 0)
+	ROM_REGION( 0x200000, "rotatingsprites:sprites_h", 0)
 	ROM_LOAD32_BYTE( "bf04d.ic30", 0x000000, 0x40000, CRC(6203d320) SHA1(d58225d8a362971a0eb63c94abc1e8c76198fd2a) ) // all rev D
 	ROM_LOAD32_BYTE( "bf12d.ic49", 0x000001, 0x40000, CRC(d261dfa7) SHA1(e787901112780e9770300999722fc80aa1d7ab18) )
 	ROM_LOAD32_BYTE( "bf06d.ic32", 0x000002, 0x40000, CRC(be187c3c) SHA1(46383eb40c0caeb1bc636630a4d849aa2d1a12d2) )
@@ -1528,7 +1544,7 @@ ROM_START( bigfightj ) // ABA-011 main board + ABA-012 daughter board
 	ROM_LOAD32_BYTE( "bf13d.ic50", 0x100001, 0x20000, CRC(1e46cd79) SHA1(c81c96b287a6cc91d3ab4dd8043153814560be3d) )
 	ROM_LOAD32_BYTE( "bf07d.ic33", 0x100002, 0x20000, CRC(4940b0bb) SHA1(762f21055921093349ca09c35ef516bde6330aa8) )
 	ROM_LOAD32_BYTE( "bf15d.ic52", 0x100003, 0x20000, CRC(dab0c80a) SHA1(a172937c9599acbd77dcac02ea7e43f576d66d8c) )
-	ROM_COPY("sprites_h",          0x100000, 0x180000, 0x080000 )
+	ROM_COPY("rotatingsprites:sprites_h",          0x100000, 0x180000, 0x080000 )
 
 	ROM_REGION( 0x20000, "cw_tileclut", 0 )
 	ROM_LOAD( "bf27.ic128", 0x000000, 0x20000, CRC(da027dcf) SHA1(47d18a8a273fea72cb3ad3d58166fe38ca28a860) )
@@ -1545,12 +1561,14 @@ ROM_END
 
 /***************************************************************************/
 
+void tatsumi_state::init_tatsumi()
+{
+	tatsumi_reset();
+}
+
 void apache3_state::init_apache3()
 {
-	m_sprite_palette_base = 0;
-	m_rom_clut_size = 0x800;
-	m_rom_clut_offset = 0x100000 - m_rom_clut_size;
-	tatsumi_reset();
+	init_tatsumi();
 
 	m_apache3_rot_idx = 0;
 
@@ -1560,29 +1578,7 @@ void apache3_state::init_apache3()
 	// TODO: ym2151_set_port_write_handler for CT1/CT2 outputs
 }
 
-void roundup5_state::init_roundup5()
-{
-	m_sprite_palette_base = 512;
-	m_rom_clut_size = 0x800;
-	m_rom_clut_offset = 0xc0000 - m_rom_clut_size;
-	tatsumi_reset();
-}
 
-void cyclwarr_state::init_cyclwarr()
-{
-	m_sprite_palette_base = 4096;
-	m_rom_clut_size = 0x1000;
-	m_rom_clut_offset = 0x100000 - m_rom_clut_size;
-	tatsumi_reset();
-}
-
-void cyclwarr_state::init_bigfight()
-{
-	m_sprite_palette_base = 4096;
-	m_rom_clut_size = 0x1000;
-	m_rom_clut_offset = 0x200000 - m_rom_clut_size;
-	tatsumi_reset();
-}
 
 /***************************************************************************/
 
@@ -1594,11 +1590,11 @@ GAME( 1988, apache3,   0,        apache3,   apache3,  apache3_state,  init_apach
 GAME( 1988, apache3a,  apache3,  apache3,   apache3,  apache3_state,  init_apache3,  ROT0, "Tatsumi", "Apache 3 (rev E)", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Rev C & E CPU code
 GAME( 1988, apache3b,  apache3,  apache3,   apache3,  apache3_state,  init_apache3,  ROT0, "Tatsumi (Kana Corporation license)", "Apache 3 (Kana Corporation license, rev G)", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Rev C & G CPU code
 
-GAMEL(1989, roundup5,  0,        roundup5,  roundup5, roundup5_state, init_roundup5, ROT0, "Tatsumi", "Round Up 5 - Super Delta Force", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING, layout_roundup5 )
+GAMEL(1989, roundup5,  0,        roundup5,  roundup5, roundup5_state, init_tatsumi, ROT0, "Tatsumi", "Round Up 5 - Super Delta Force", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING, layout_roundup5 )
 
-GAME( 1991, cyclwarr,  0,        cyclwarr,  cyclwarr, cyclwarr_state, init_cyclwarr, ROT0, "Tatsumi", "Cycle Warriors (rev C)", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Rev C & B CPU code
-GAME( 1991, cyclwarra, cyclwarr, cyclwarr,  cyclwarb, cyclwarr_state, init_cyclwarr, ROT0, "Tatsumi", "Cycle Warriors (rev B)", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Rev B & A CPU code
-GAME( 1991, cyclwarrb, cyclwarr, cyclwarr,  cyclwarb, cyclwarr_state, init_cyclwarr, ROT0, "Tatsumi", "Cycle Warriors", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Original version with no Rev roms
+GAME( 1991, cyclwarr,  0,        cyclwarr,  cyclwarr, cyclwarr_state, init_tatsumi, ROT0, "Tatsumi", "Cycle Warriors (rev C)", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Rev C & B CPU code
+GAME( 1991, cyclwarra, cyclwarr, cyclwarr,  cyclwarb, cyclwarr_state, init_tatsumi, ROT0, "Tatsumi", "Cycle Warriors (rev B)", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Rev B & A CPU code
+GAME( 1991, cyclwarrb, cyclwarr, cyclwarr,  cyclwarb, cyclwarr_state, init_tatsumi, ROT0, "Tatsumi", "Cycle Warriors", MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // Original version with no Rev roms
 
-GAME( 1992, bigfight,  0,        bigfight,  bigfight, cyclwarr_state, init_bigfight, ROT0, "Tatsumi", "Big Fight - Big Trouble In The Atlantic Ocean", MACHINE_IMPERFECT_GRAPHICS )
-GAME( 1992, bigfightj, bigfight, bigfight,  bigfight, cyclwarr_state, init_bigfight, ROT0, "Tatsumi", "Big Fight - Big Trouble In The Atlantic Ocean (Japan, rev F)", MACHINE_IMPERFECT_GRAPHICS ) // Rev D through F CPU codes
+GAME( 1992, bigfight,  0,        bigfight,  bigfight, cyclwarr_state, init_tatsumi, ROT0, "Tatsumi", "Big Fight - Big Trouble In The Atlantic Ocean", MACHINE_IMPERFECT_GRAPHICS )
+GAME( 1992, bigfightj, bigfight, bigfight,  bigfight, cyclwarr_state, init_tatsumi, ROT0, "Tatsumi", "Big Fight - Big Trouble In The Atlantic Ocean (Japan, rev F)", MACHINE_IMPERFECT_GRAPHICS ) // Rev D through F CPU codes
